requirements.txt
streamlit
pandas
plotly
qrcode
pillow
folium
streamlit-folium
bcrypt
psycopg2-binary
